+++ SSDDriver/SSDDriver.vcxproj
@@ -132,7 +132,11 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="fileIo_test.cpp" />
+    <ClCompile Include="file_io.cpp" />
     <ClCompile Include="main.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="file_io.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
