+++ SSDDriver/ssd_nand.txt
@@ -1,2 +1,6 @@
+2 0xAAAABBBB
+2 0xAAAABBBB
 3 0x1298CDEF
 2 0xAAAABBBB
+2 0xAAAABBBB
+3 0x1298CDEF
